Split out controllers from views.py
