+++ mathics_django/web/__init__.py
@@ -1,2 +1,4 @@
-#!/usr/bin/env python3
 # -*- coding: utf-8 -*-
+"""
+Python-related web code.
+"""
