+++ mathics_django/web/controllers/__init__.py
@@ -1,0 +1,3 @@
+"""
+Modules for each Web Page section.
+"""
